+++ HISTORY.rst
@@ -2,6 +2,15 @@
 History
 =======
 
+0.32.0 (2024-XX-XX)
+-------------------
+
+* Add Leiden clustering algorithm
+* Add functions to detect and break cycles
+* Add damping factor in diffusion
+* Fix F1 scores
+* Get clustering coefficient for directed graphs
+
 
 0.31.0 (2023-05-22)
 -------------------